feat: Remove background_task and hide get_current_transaction.
background_task is seemingly unusued, so it has been removed.

Also add comments explaining get_current_transaction is
internal-only and will not be exposed via API. Remove most
public references to it. Also refactor code so that it
is only used in the file where it's used.
